Change model and controller directory and add into v1 modules
