+++ modules/v1/users/UserModule.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace app\modules\v1\users;
+
+
+use yii\base\Module;
+
+
+/**
+ * Class UserModule
+ *
+ * @package app\modules\v1\users
+ */
+class UserModule extends Module
+{
+
+}
