+++ modules/v1/users/resources/UserProfileResource.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace app\modules\v1\setup\resources;
+namespace app\modules\v1\users\resources;
 
 
 use app\models\UserProfile;
@@ -9,7 +9,7 @@ use app\models\UserProfile;
 /**
  * Class UserProfileResource
  *
- * @package app\modules\v1\setup\resources
+ * @package app\modules\v1\users\resources
  */
 class UserProfileResource extends UserProfile
 {
